% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_blocos_paises.R
\name{comerciobr_blocos_paises}
\alias{comerciobr_blocos_paises}
\title{Lista de pa\u00edses a partir de um bloco}
\usage{
comerciobr_blocos_paises(bloco)
}
\arguments{
\item{bloco}{o nome de um bloco, de acordo com a base dic_blocos do pacote "comerciobr"}
}
\value{
lista de pa\u00edses associados a determinado bloco
}
\description{
Lista de pa\u00edses a partir de um bloco
}
