% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_dados_concentracao.R
\name{comerciobr_dados_concentracao}
\alias{comerciobr_dados_concentracao}
\title{Dados de concentração}
\usage{
comerciobr_dados_concentracao(pais, periodo, fluxo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Dados de concentração
}
