% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_dados_intraindustria.R
\name{comerciobr_dados_intraindustria}
\alias{comerciobr_dados_intraindustria}
\title{Dados Intraindustria}
\usage{
comerciobr_dados_intraindustria(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Dados Intraindustria
}
