% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_dados_paises.R
\name{comerciobr_dados_paises}
\alias{comerciobr_dados_paises}
\title{Dados de com\u00e9rcio bilateral de um pa\u00eds em rela\u00e7\u00e3o a seus semelhantes}
\usage{
comerciobr_dados_paises(pais, periodo)
}
\arguments{
\item{pais}{um pa\u00eds}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Dados de com\u00e9rcio do Brasil com um determinado pa\u00eds, comparado com pa\u00edses
em condi\u00e7\u00f5es semelhantes (selecionados pelo ranking).
}
