% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_get_ultimoano.R
\name{comerciobr_get_ulimoano}
\alias{comerciobr_get_ulimoano}
\title{\u00daltimo ano da base de dados}
\usage{
comerciobr_get_ulimoano()
}
\value{
o \u00faltimo ano dispon\u00edvel na base de dados "comerciobr"
}
\description{
\u00daltimo ano da base de dados
}
