% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/comerciobr_grafico_concentracao_intraindustria.R
\name{comerciobr_grafico_concentracao_intraindustria}
\alias{comerciobr_grafico_concentracao_intraindustria}
\title{Gráfico de índice de Concentração de Comércio - Índice de Herfindahl-Hirschman (HH) e Índice Intraindústria - Índice de Grubel-Lloyd (GG)}
\usage{
comerciobr_grafico_concentracao_intraindustria(periodo, pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico de índice de Concentração de Comércio - Índice de Herfindahl-Hirschman (HH) e Índice Intraindústria - Índice de Grubel-Lloyd (GG)
}
