% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_corrente.R
\name{comerciobr_grafico_corrente}
\alias{comerciobr_grafico_corrente}
\title{Gr\u00e1fico de Fluxo de Com\u00e9rcio Brasil-pa\u00eds}
\usage{
comerciobr_grafico_corrente(pais, periodo)
}
\arguments{
\item{pais}{um pa\u00eds}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Com base na fun\u00e7\u00e3o \code{comerciobr_dados_corrente}.
}
