% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_paises.R
\name{comerciobr_grafico_paises}
\alias{comerciobr_grafico_paises}
\title{Gr\u00e1fico de compara\u00e7\u00e3o do fluxo de com\u00e9rcio de um pa\u00eds com o Brasil, em rela\u00e7\u00e3o a pa\u00edses semelhantes}
\usage{
comerciobr_grafico_paises(pais, periodo)
}
\arguments{
\item{pais}{um pa\u00eds}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gr\u00e1fico de compara\u00e7\u00e3o do fluxo de com\u00e9rcio de um pa\u00eds com o Brasil, em rela\u00e7\u00e3o a pa\u00edses semelhantes
}
