% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_paises_proporcao.R
\name{comerciobr_grafico_paises_proporcao}
\alias{comerciobr_grafico_paises_proporcao}
\title{Gr\u00e1fico de ranking e propor\u00e7\u00e3o do fluxo de com\u00e9rcio do brasil com um determinado pa\u00eds}
\usage{
comerciobr_grafico_paises_proporcao(pais, periodo)
}
\arguments{
\item{pais}{um pa\u00eds}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gr\u00e1fico de ranking e propor\u00e7\u00e3o do fluxo de com\u00e9rcio do brasil com um determinado pa\u00eds
}
