% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_produtos.R
\name{comerciobr_grafico_produtos}
\alias{comerciobr_grafico_produtos}
\title{Gr\u00e1fico dos cinco principais produtos comercializados do Brasil com um pa\u00eds}
\usage{
comerciobr_grafico_produtos(pais, periodo)
}
\arguments{
\item{pais}{um pa\u00eds}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gr\u00e1fico dos cinco principais produtos comercializados do Brasil com um pa\u00eds
}
