% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_lista_lista.R
\name{comerciobr_lista_lista}
\alias{comerciobr_lista_lista}
\title{Lista de países a partir de um bloco}
\usage{
comerciobr_lista_lista(x)
}
\arguments{
\item{bloco}{o nome de um bloco, de acordo com a base dic_blocos do pacote "comerciobr2"}
}
\value{
lista de países associados a determinado bloco
}
\description{
Lista de países a partir de um bloco
}
