% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_tabelao_corrente.R
\name{comerciobr_tabelao_corrente}
\alias{comerciobr_tabelao_corrente}
\title{Tabelao da corrente de comércio do Brasil com um determinado país}
\usage{
comerciobr_tabelao_corrente(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Tabelao da corrente de comércio do Brasil com um determinado país
}
