% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_tabelao_produtos.R
\name{comerciobr_tabelao_produtos}
\alias{comerciobr_tabelao_produtos}
\title{Tabelao de principais produtos exportados e importados pelo Brasil a um determinado pa\\\\u00eds}
\usage{
comerciobr_tabelao_produtos(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Tabelao de principais produtos exportados e importados pelo Brasil a um determinado pa\\\\u00eds
}
