% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_tabelas_comercio.R
\name{comerciobr_tabelas_comercio}
\alias{comerciobr_tabelas_comercio}
\title{Tabela Intraindustria}
\usage{
comerciobr_tabelas_comercio(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Tabela Intraindustria
}
