% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_dados_paises.R
\name{comerciomundo_dados_paises}
\alias{comerciomundo_dados_paises}
\title{Dados de com\u00e9rcio um pa\u00eds com todos os demais}
\usage{
comerciomundo_dados_paises(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\description{
Dados de com\u00e9rcio um pa\u00eds com todos os demais
}
