% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_get_ultimoano.R
\name{comerciomundo_get_ultimoano}
\alias{comerciomundo_get_ultimoano}
\title{\u00daltimo ano da base de dados}
\usage{
comerciomundo_get_ultimoano(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\value{
o \u00faltimo ano dispon\u00edvel na base de dados "comerciomundo"
}
\description{
\u00daltimo ano da base de dados
}
