% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_corrente.R
\name{comerciomundo_grafico_corrente}
\alias{comerciomundo_grafico_corrente}
\title{Gr\u00e1fico de Fluxo de Com\u00e9rcio pa\u00eds-mundo}
\usage{
comerciomundo_grafico_corrente(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\description{
Com base na fun\u00e7\u00e3o \code{comerciomundo_dados_corrente}.
}
