% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_paises_ranking.R
\name{comerciomundo_grafico_paises_ranking}
\alias{comerciomundo_grafico_paises_ranking}
\title{Gr\u00e1fico de compara\u00e7\u00e3o do fluxo de com\u00e9rcio de um pa\u00eds com seus 10 principais parceiros}
\usage{
comerciomundo_grafico_paises_ranking(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\description{
Gr\u00e1fico de compara\u00e7\u00e3o do fluxo de com\u00e9rcio de um pa\u00eds com seus 10 principais parceiros
}
