% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_report_pais.R
\name{comerciomundo_report_pais}
\alias{comerciomundo_report_pais}
\title{Relat\u00f3rio de rela\u00e7\u00f5es bilaterais comerciais pa\u00eds-mundo}
\usage{
comerciomundo_report_pais(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\description{
Relat\u00f3rio de rela\u00e7\u00f5es bilaterais comerciais pa\u00eds-mundo
}
