% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bloco.R
\name{get_bloco}
\alias{get_bloco}
\title{Fun\u00e7\u00e3o que devolve um bloco associado a um pa\u00eds (ou lista de pa\u00edses)}
\usage{
get_bloco(paises)
}
\arguments{
\item{paises}{um pa\u00eds ou vetor de pa\u00edses}
}
\description{
Fun\u00e7\u00e3o que devolve um bloco associado a um pa\u00eds (ou lista de pa\u00edses)
}
