% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pais.R
\name{get_pais}
\alias{get_pais}
\title{Normaliza nomes dos países em diferentes bases de dados}
\usage{
get_pais(pais, base)
}
\arguments{
\item{pais}{nome do país desejado (seja em português ou em inglês)}

\item{base}{base de dados de referência (pode ser "fmi", "mecon" ou "comtrade")}
}
\value{
retorna o nome do país como adequado para a base em referência.
}
\description{
Normaliza nomes dos países em diferentes bases de dados
}
