% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meses.R
\name{meses}
\alias{meses}
\title{Fun\u00e7\u00e3o que devolve meses em portugu\u00eas}
\usage{
meses(num)
}
\arguments{
\item{num}{n\u00famero de um m\u00eas}
}
\description{
Fun\u00e7\u00e3o que devolve meses em portugu\u00eas
}
