% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_report_pais.R
\name{relatorio_barao}
\alias{relatorio_barao}
\title{Relatório de relações bilaterais comerciais Brasil-país}
\usage{
relatorio_barao(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Relatório de relações bilaterais comerciais Brasil-país
}
