% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_brasil_bloco.R
\name{relatorio_brasil_bloco}
\alias{relatorio_brasil_bloco}
\title{Relatório de relações bilaterais comerciais Brasil-país}
\usage{
relatorio_brasil_bloco(bloco)
}
\arguments{
\item{pais}{um país}
}
\description{
Relatório de relações bilaterais comerciais Brasil-país
}
