% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_brasil_pais.R
\name{relatorio_brasil_pais}
\alias{relatorio_brasil_pais}
\title{Relat\u00f3rio de rela\u00e7\u00f5es bilaterais comerciais Brasil-pa\u00eds}
\usage{
relatorio_brasil_pais(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\description{
Relat\u00f3rio de rela\u00e7\u00f5es bilaterais comerciais Brasil-pa\u00eds
}
