% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_brasil_pais_tabelao.R
\name{relatorio_brasil_pais_tabelao}
\alias{relatorio_brasil_pais_tabelao}
\title{Relatório de relações bilaterais comerciais Brasil-país em formato
tabelão}
\usage{
relatorio_brasil_pais_tabelao(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Relatório de relações bilaterais comerciais Brasil-país em formato
tabelão
}
