% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_pais_mundo.R
\name{relatorio_pais_mundo}
\alias{relatorio_pais_mundo}
\title{Relat\u00f3rio de rela\u00e7\u00f5es bilaterais comerciais Pa\u00eds-Mundo}
\usage{
relatorio_pais_mundo(pais)
}
\arguments{
\item{pais}{um pa\u00eds}
}
\description{
Relat\u00f3rio de rela\u00e7\u00f5es bilaterais comerciais Pa\u00eds-Mundo
}
