% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_itamaraty.R
\name{template_itamaraty}
\alias{template_itamaraty}
\title{Template de relatório para o Itamaraty}
\usage{
template_itamaraty(file)
}
\arguments{
\item{file}{nome do arquivo a ser criado}
}
\description{
Template de relatório para o Itamaraty
}
