% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_dados_corrente.R
\name{comerciobr_dados_corrente}
\alias{comerciobr_dados_corrente}
\title{Dados de fluxo de comércio bilateral entre Brasil e um determinado país}
\usage{
comerciobr_dados_corrente(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\value{
tibble com dados de comércio bilateral de 2010 em diante.
}
\description{
Dados de fluxo de comércio bilateral entre Brasil e um determinado país
}
