% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_dados_fatores.R
\name{comerciobr_dados_fatores}
\alias{comerciobr_dados_fatores}
\title{Dados de comércio por agregaçao}
\usage{
comerciobr_dados_fatores(pais, periodo, fator)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}

\item{fator}{agregaçao: "isic", "fator", "cgce" ou "cuci".}
}
\description{
Dados de comércio selecionados pela agregaçao "ISIC", "Fator Agregado", "CGCE" e "CUCI"
}
