% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_dados_paises.R
\name{comerciobr_dados_paises}
\alias{comerciobr_dados_paises}
\title{Dados de comércio bilateral de um país em relação a seus semelhantes}
\usage{
comerciobr_dados_paises(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Dados de comércio do Brasil com um determinado país, comparado com países
em condições semelhantes (selecionados pelo ranking).
}
