% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_get_ultimoano.R
\name{comerciobr_get_ulimoano}
\alias{comerciobr_get_ulimoano}
\title{Último ano da base de dados}
\usage{
comerciobr_get_ulimoano()
}
\value{
o último ano disponível na base de dados "comerciobr"
}
\description{
Último ano da base de dados
}
