% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_get_ultimomes.R
\name{comerciobr_get_ultimomes}
\alias{comerciobr_get_ultimomes}
\title{Último mês disponível do último ano disponível em comerciobr}
\usage{
comerciobr_get_ultimomes()
}
\description{
Último mês disponível do último ano disponível em comerciobr
}
