% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_corrente.R
\name{comerciobr_grafico_corrente}
\alias{comerciobr_grafico_corrente}
\title{Gráfico de Fluxo de Comércio Brasil-país}
\usage{
comerciobr_grafico_corrente(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Com base na função \code{comerciobr_dados_corrente}.
}
