% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_fatores.R
\name{comerciobr_grafico_fatores}
\alias{comerciobr_grafico_fatores}
\title{Gráfico de Composição de Comércio por Agregação}
\usage{
comerciobr_grafico_fatores(pais, periodo, fator)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}

\item{fator}{"isic", "fator", "cgce", "cuci"}
}
\description{
Gráfico de Composição de Comércio por Agregação
}
