% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_paises.R
\name{comerciobr_grafico_paises}
\alias{comerciobr_grafico_paises}
\title{Gráfico de comparação do fluxo de comércio de um país com o Brasil,
em relação a países semelhantes}
\usage{
comerciobr_grafico_paises(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gráfico de comparação do fluxo de comércio de um país com o Brasil,
em relação a países semelhantes
}
