% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_paises_evolucao.R
\name{comerciobr_grafico_paises_evolucao}
\alias{comerciobr_grafico_paises_evolucao}
\title{Gráfico de evolução do fluxo de comércio de um país com o Brasil, comparado com países semelhantes}
\usage{
comerciobr_grafico_paises_evolucao(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gráfico de evolução do fluxo de comércio de um país com o Brasil, comparado com países semelhantes
}
