% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_paises_proporcao.R
\name{comerciobr_grafico_paises_proporcao}
\alias{comerciobr_grafico_paises_proporcao}
\title{Gráfico de ranking e proporção do fluxo de comércio do brasil com um determinado país}
\usage{
comerciobr_grafico_paises_proporcao(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gráfico de ranking e proporção do fluxo de comércio do brasil com um determinado país
}
