% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_produtos.R
\name{comerciobr_grafico_produtos}
\alias{comerciobr_grafico_produtos}
\title{Gráfico dos cinco principais produtos comercializados do Brasil com um país}
\usage{
comerciobr_grafico_produtos(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gráfico dos cinco principais produtos comercializados do Brasil com um país
}
