% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_grafico_produtos_proporcao.R
\name{comerciobr_grafico_produtos_proporcao}
\alias{comerciobr_grafico_produtos_proporcao}
\title{Gráfico de proporção dos produtos comercializados do Brasil com um país}
\usage{
comerciobr_grafico_produtos_proporcao(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Gráfico de proporção dos produtos comercializados do Brasil com um país
}
