% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_tabela_fatores.R
\name{comerciobr_tabela_fatores}
\alias{comerciobr_tabela_fatores}
\title{Tabela de dados de comércio por níveis de agregação}
\usage{
comerciobr_tabela_fatores(pais, periodo, fator)
}
\arguments{
\item{pais}{um países}

\item{periodo}{"anual" ou "mensal"}

\item{fator}{"isic", "fator", "cgce" ou "cuci"}
}
\description{
Tabela de dados de comércio por níveis de agregação
}
