% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciobr_tabela_paises.R
\name{comerciobr_tabela_paises}
\alias{comerciobr_tabela_paises}
\title{Tabela de fluxo comercial por país}
\usage{
comerciobr_tabela_paises(pais, periodo)
}
\arguments{
\item{pais}{um país}

\item{periodo}{"anual" ou "mensal"}
}
\description{
Tabela de fluxo comercial por país
}
