% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_dados_corrente.R
\name{comerciomundo_dados_corrente}
\alias{comerciomundo_dados_corrente}
\title{Dados de fluxo de comércio bilateral entre o países alvo e o mundo}
\usage{
comerciomundo_dados_corrente(pais)
}
\arguments{
\item{pais}{um países}
}
\description{
Dados de fluxo de comércio bilateral entre o países alvo e o mundo
}
