% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_dados_paises.R
\name{comerciomundo_dados_paises}
\alias{comerciomundo_dados_paises}
\title{Dados de comércio de um país com todos os demais}
\usage{
comerciomundo_dados_paises(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Dados de comércio de um país com todos os demais
}
