% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_dados_produtos.R
\name{comerciomundo_dados_produtos}
\alias{comerciomundo_dados_produtos}
\title{Dados dos principais produtos comercializados de um país com o mundo}
\usage{
comerciomundo_dados_produtos(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Dados dos principais produtos comercializados de um país com o mundo
}
