% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_get_ultimoano.R
\name{comerciomundo_get_ultimoano}
\alias{comerciomundo_get_ultimoano}
\title{Último ano da base de dados}
\usage{
comerciomundo_get_ultimoano(pais)
}
\arguments{
\item{pais}{um país}
}
\value{
o último ano disponível na base de dados "comerciomundo"
}
\description{
Último ano da base de dados
}
