% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_corrente.R
\name{comerciomundo_grafico_corrente}
\alias{comerciomundo_grafico_corrente}
\title{Gráfico de Fluxo de Comércio país-mundo}
\usage{
comerciomundo_grafico_corrente(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Com base na função \code{comerciomundo_dados_corrente}.
}
