% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_paises_evolucao.R
\name{comerciomundo_grafico_paises_evolucao}
\alias{comerciomundo_grafico_paises_evolucao}
\title{Gráfico de evolução do fluxo de comércio de um país com seus cinco principais parceiros}
\usage{
comerciomundo_grafico_paises_evolucao(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico de evolução do fluxo de comércio de um país com seus cinco principais parceiros
}
