% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_paises_proporcao.R
\name{comerciomundo_grafico_paises_proporcao}
\alias{comerciomundo_grafico_paises_proporcao}
\title{Gráfico de ranking e proporção do fluxo de comércio do país com os demais país}
\usage{
comerciomundo_grafico_paises_proporcao(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico de ranking e proporção do fluxo de comércio do país com os demais país
}
