% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_paises_ranking.R
\name{comerciomundo_grafico_paises_ranking}
\alias{comerciomundo_grafico_paises_ranking}
\title{Gráfico de comparação do fluxo de comércio de um país com seus 10 principais parceiros}
\usage{
comerciomundo_grafico_paises_ranking(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico de comparação do fluxo de comércio de um país com seus 10 principais parceiros
}
