% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_produtos_evolucao.R
\name{comerciomundo_grafico_produtos_evolucao}
\alias{comerciomundo_grafico_produtos_evolucao}
\title{Gráfico da evolução dos principais produtos comercializados por um país com o mundo no último ano disponível}
\usage{
comerciomundo_grafico_produtos_evolucao(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico da evolução dos principais produtos comercializados por um país com o mundo no último ano disponível
}
