% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_produtos_proporcao.R
\name{comerciomundo_grafico_produtos_proporcao}
\alias{comerciomundo_grafico_produtos_proporcao}
\title{Gráfico, em proporção, dos principais produtos comercializados por um país com o mundo no último ano disponível}
\usage{
comerciomundo_grafico_produtos_proporcao(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico, em proporção, dos principais produtos comercializados por um país com o mundo no último ano disponível
}
