% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_grafico_produtos_ranking.R
\name{comerciomundo_grafico_produtos_ranking}
\alias{comerciomundo_grafico_produtos_ranking}
\title{Gráfico dos principais produtos comercializados por um país no último ano disponível}
\usage{
comerciomundo_grafico_produtos_ranking(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Gráfico dos principais produtos comercializados por um país no último ano disponível
}
