% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_report_pais.R
\name{comerciomundo_report_pais}
\alias{comerciomundo_report_pais}
\title{Relatório de relações bilaterais comerciais país-mundo}
\usage{
comerciomundo_report_pais(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Relatório de relações bilaterais comerciais país-mundo
}
