% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comerciomundo_tabela_corrente.R
\name{comerciomundo_tabela_corrente}
\alias{comerciomundo_tabela_corrente}
\title{Tabela de Corrente de Comércio país-mundo}
\usage{
comerciomundo_tabela_corrente(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
em formato latex.
}
