% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic_paises_fmi}
\alias{dic_paises_fmi}
\title{Dicionário de Países - MEcon/FMI}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 239 rows and 2 columns.
}
\usage{
dic_paises_fmi
}
\description{
Nomes dos países em inglês, na grafia do FMI e, em português, na grafia do Ministério da Economia.
}
\keyword{datasets}
