% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bloco.R
\name{get_bloco}
\alias{get_bloco}
\title{Função que devolve um bloco associado a um país (ou lista de países)}
\usage{
get_bloco(paises)
}
\arguments{
\item{paises}{um paíes ou vetor de países}
}
\description{
Função que devolve um bloco associado a um país (ou lista de países)
}
