% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_brasil_pais.R
\name{relatorio_brasil_pais}
\alias{relatorio_brasil_pais}
\title{Relatório de relações bilaterais comerciais Brasil-país}
\usage{
relatorio_brasil_pais(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
Relatório de relações bilaterais comerciais Brasil-país
}
