% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_pais_mundo.R
\name{relatorio_pais_mundo}
\alias{relatorio_pais_mundo}
\title{Relatório de relações bilaterais comerciais País-Mundo}
\usage{
relatorio_pais_mundo(pais)
}
\arguments{
\item{pais}{um país}
}
\description{
O relatório deve ser usado em conjunto com o objeto "dic_comtrade_mdic",
do pacote comerciomundo .
}
